fix: add asset-loader file that can house the asset location logic
